Readme.md
Reclassification of the raster layer using a table of range values ​​extracted directly from the raster band, or defined directly by the user. The histogram of the raster is shown, reflecting the distribution of values ​​within the minimum and maximum range of the selected reference band. To run the plugin you need to install the following libraries numpy, matplotlib, gdal, csv and locale.